backend/requirements.txt
# Core Django
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1

# API Documentation
drf-spectacular==0.27.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2

# Environment & Configuration
python-decouple==3.8
python-magic==0.4.27

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Rate Limiting
django-ratelimit==4.1.0
